requirements.txt
requests>=2.31.0
feedparser>=6.0.11
beautifulsoup4>=4.12.2
python-dateutil>=2.9.0
rapidfuzz>=3.7.0
pyyaml>=6.0.1
google-genai>=0.3.0
